Fix missing inline for pika::mpi::experimental::detail::error_message
An inline keyword would fix it, but this moves the function into a
source file and PIKA_EXPORTs the function instead.
